refactor task editor bottom sheet
