+++ app/src/main/java/com/giraffe/tudeeapp/di/appModule.kt
@@ -10,7 +10,7 @@ import com.giraffe.tudeeapp.domain.service.SplashService
 import com.giraffe.tudeeapp.domain.service.TasksService
 import com.giraffe.tudeeapp.presentation.shared.MainViewModel
 import com.giraffe.tudeeapp.presentation.home.HomeViewModel
-import com.giraffe.tudeeapp.presentation.shared.addedittask.TaskEditorViewModel
+import com.giraffe.tudeeapp.presentation.shared.taskeditor.TaskEditorViewModel
 import com.giraffe.tudeeapp.presentation.shared.taskdetails.TaskDetailsViewModel
 import com.giraffe.tudeeapp.presentation.splash.viewmodel.SplashViewModel
 import org.koin.android.ext.koin.androidApplication
