--- app/src/main/java/com/giraffe/tudeeapp/presentation/shared/addedittask/TaskEditorBottomSheet.kt
@@ -1,89 +1,0 @@
-package com.giraffe.tudeeapp.presentation.shared.addedittask
-
-
-import android.util.Log
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.unit.dp
-import com.giraffe.tudeeapp.design_system.theme.Theme
-import kotlinx.coroutines.launch
-import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun TaskEditorBottomSheet(
-    taskId: Long?,
-    onDismissRequest: () -> Unit,
-    headerTitle: String,
-    saveButtonText: String,
-    modifier: Modifier = Modifier,
-    onSuccessSave: () -> Unit = {},
-    onError: (String?) -> Unit = {},
-    viewModel: TaskEditorViewModel = koinViewModel { parametersOf(taskId) },
-) {
-    val screenHeight = LocalConfiguration.current.screenHeightDp.dp
-    val maxSheetHeight = screenHeight - 72.dp
-    val taskState by viewModel.taskState.collectAsState()
-
-    val scope = rememberCoroutineScope()
-    val bottomSheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-
-    if (taskState.isSuccessSave) {
-        Log.d("TAG", "TaskEditorBottomSheet: 1111111")
-        onSuccessSave()
-    }
-
-    if (taskState.errorMessage != null) {
-        onError(taskState.errorMessage)
-    }
-
-    ModalBottomSheet(
-        onDismissRequest = {
-            viewModel.onCancel()
-            onDismissRequest()
-            scope.launch { bottomSheetState.hide() }
-        },
-        sheetState = bottomSheetState,
-        modifier = modifier
-            .fillMaxWidth()
-            .heightIn(max = maxSheetHeight),
-        containerColor = Theme.color.surface
-    ) {
-        TaskEditorBottomSheetContent(
-            headerTitle = headerTitle,
-            saveButtonText = saveButtonText,
-            taskState = taskState,
-            categories = taskState.categories,
-            isLoading = taskState.isLoading,
-            categoriesLoading = taskState.isLoading,
-            onTitleChange = { viewModel.onTitleChange(it) },
-            onDescriptionChange = { viewModel.onDescriptionChange(it) },
-            onPriorityChange = { viewModel.onPriorityChange(it) },
-            onCategoryChange = { viewModel.onCategoryChange(it) },
-            onDueDateChange = viewModel::onDueDateChange
-            ,
-            onSaveClick = {
-                viewModel.saveTask()
-                onDismissRequest()
-            },
-            onCancelClick = {
-                viewModel.onCancel()
-                scope.launch { bottomSheetState.hide() }
-                onDismissRequest()
-            }
-        )
-    }
-
-
-}
-
