+++ app/src/main/java/com/giraffe/tudeeapp/presentation/shared/taskeditor/TaskEditorUiState.kt
@@ -5,11 +5,11 @@ import com.giraffe.tudeeapp.domain.util.DomainError
 import com.giraffe.tudeeapp.presentation.uimodel.TaskUi
 
 data class TaskEditorUiState(
-    val taskUi: TaskUi? = null,
+    val taskUi: TaskUi = TaskUi(),
     val categories: List<Category> = emptyList(),
     val isLoading: Boolean = false,
     val isSuccessAdded: Boolean = false,
     val isSuccessEdited: Boolean = false,
-    val isValidInput: Boolean = false
+    val isValidTask: Boolean = false,
 )
