+++ app/src/main/java/com/giraffe/tudeeapp/presentation/uimodel/TaskUi.kt
@@ -4,18 +4,25 @@ import com.giraffe.tudeeapp.domain.model.Category
 import com.giraffe.tudeeapp.domain.model.task.Task
 import com.giraffe.tudeeapp.domain.model.task.TaskPriority
 import com.giraffe.tudeeapp.domain.model.task.TaskStatus
+import com.giraffe.tudeeapp.presentation.utils.getCurrentLocalDateTime
 import kotlinx.datetime.LocalDateTime
 
 data class TaskUi (
-    val id: Long,
-    val title: String,
-    val description: String,
-    val priorityType: TaskPriority,
-    val status: TaskStatus,
-    val category: Category,
-    val dueDate: LocalDateTime,
-    val createdAt: LocalDateTime,
-    val updatedAt: LocalDateTime
+    val id: Long = 0L,
+    val title: String = "",
+    val description: String = "",
+    val priorityType: TaskPriority = TaskPriority.LOW,
+    val status: TaskStatus = TaskStatus.TODO,
+    val category: Category = Category(
+        id = 0L,
+        name = "",
+        imageUri = "",
+        isEditable = true,
+        taskCount = 0
+    ),
+    val dueDate: LocalDateTime = getCurrentLocalDateTime(),
+    val createdAt: LocalDateTime = getCurrentLocalDateTime(),
+    val updatedAt: LocalDateTime = getCurrentLocalDateTime()
 )
 
 fun Task.toTaskUi(category: Category) = TaskUi(
