+++ app/src/main/java/com/giraffe/tudeeapp/presentation/utils/getCurrentLocalDateTime.kt
@@ -1,0 +1,10 @@
+package com.giraffe.tudeeapp.presentation.utils
+
+import kotlinx.datetime.Clock
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.toLocalDateTime
+
+fun getCurrentLocalDateTime(): LocalDateTime {
+    return Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault())
+}
